Outline the codec for the wire protocol
The codec uses the Rust type system to restrict the which messages it
can encode or decode depending on the role it is playing (server or
client) and the protocol family it is designed to handle. The codec will
not encode messages that are not a part of the protocol family. A codec
in the server role will not encode request messages and in the client
role it will not encode event messages (attempts will fail to compile).

This commit also has the implementation of the writer for the sever
argument types.