fix: change MSRV and big-endian platform for ci
cargo-build has an indirect dependancy on the xml-rs crate which has
an effective MSRV of 1.40.0 (becuase of its use of #[cfg(doctest)]).
Increase the MSRV of the project to 1.40.0 as a result and fix the one
use of code that required a later MSRV. Change the big-endian platform
used for CI to one that is still supported by the cross tool.

BREAKING CHANGE: increase the MSRV to 1.40.0