refactor(core): simply trait bound for codec and transport
Extact the complicated trait bounds for the message types for
WaylandCodec and WaylandTransport into new traits with blanket
impl: EventMessage, RequestMessage, and ProtocolFamilyMessage. These
traits are designed to be used as trait bounds and not to be implemented
directly. The only implementation of these traits should be the provided
blanket implementation.