feat(core): add functions to make MessageList entries
The from_signature() associated function in a Message allows the
creation of that kind of Message from its ObjectId and its arguments.
This is used by the MessageMaker implementation on
transport::DispatchMessage to create a Message once we know the
concreate Message type.

The from_opcode() associated function in a MessageList allows the
creation of an entry of the MessageList (including its contained
Message) given an OpCode and a MessageMaker. In practice the
MessageMaker will be a transport::DispatchMessage.

These functions will be used as a part of dispatching incoming
messages.