refactor(core): change the Error support library
The old Error support library was thiserror. Change to snafu to
support better patterns for error handling.

BREAKING CHANGE: change FromOpcodeError variants to struct variants

The FromOpcodeError variants had been tuple variants but have been
changed to struct variants as a part of a change to the internal
Error support library. Most uses of FromOpcodeError are in code
generated by caledon-build so will not be affected. Handwritten
code that directly constructs a FromOpcodeError will have to be
changed to struct variant notation and will have to name the
fields to be initialized.