feat(core): add requests and events lists to ProtocolFamily
The lists are a type that implements the new ProtocolFamilyMessageList
trait and are generated from the Waland protocol XML files. These types
(which are an emum in the generated code) give a single type for all of
the requests in the ProtocolFamily (and similarly for the events). This
type is intended as the items for a Sink implementation.