fix(core): unify the item type for transport sink
Change the item type for the WaylandTransport Sink and the WaylandCodec
Encoder to be one of the ProtocolFamilyMessageList's associated with
the ProtocolFamily (Events for the ServerRole and Requests for the
ClientRole). The old implementations of Encoder and Sink used a Message
type as the item but this would have allowed just 1 message type to be
sent over a particular transport's Sink. This change allows any request
(for ClientRole) or event (for ServerRole) associated with the ProtocolFamily
to be sent over a single transport.