feat(core): Add skeleton facade for the wire protocol
The make_wire_protocol() function and its associated types implement a
facade over the rest of the Wayland wire protocol implementation. This
is currently a skeleton (with many "todo's").

The client module uses the wire protocol facade to provide a skeleton of
the initial implementation of the client-side of the Wayland core
object-level protocol.