feat(core): partial implementation of make_wire_protocol
Add the creation of the ObjectMap and the WaylandTransport to
make_wire_protocol. This requires adjusting some of the trait bounds of
the generic types for make_wire_protocol.

This also required adding a new "type function" SendMsg (and its
associated alias SendMsgType) which map a Role (ClientRole or
ServerRole) and  ProtocolFamily to the list of messages that can be sent
using the wire protocol for the ProtocolFamily in the Role.

Change WaylandTransport and WaylandCodec to use the new SendMsg type
function instead of having separate implementations for ClientRole and
ServerRole for parts of their code (i.e. for their Sink and Encoder
implementations respectively).