feat(core): implement WireRecv including Stream
Add make_event_message() and make_request_message() to ProtocolFamily
and add a blanket trait to the role module to make a message for a given
opcode on a particular live protocol object. This is used, in
conjuction with the ObjectMap, by the Stream implementation to map from
a transport-level DispatchMessage to the appropriate protocol message.

make_event_message() and make_request_message() are currently
placeholders that will be replaced with the real implementations in a
future commit.