feat(core): implement WireState with new methods
The new methods are added to WaylandState and are designed to sport the
creation and removal of live protocol objects. Add new create(), add(),
and remove() methods to the ObjectMap to support the new WireState
methods.

These methods rely on a new IdSource type to manage the creation of new
object id's in the allowed ranges for the ClientRole or ServerRole and
in a dense manner.