fix(core): change the BitVec used in ObjectMap
The IdSource in ObjectMap uses a BitVec to track the objected id's that
are in use. Change the BitVec BitStore type from the default usize to an
explicit u64 because a 32 bit BitStore (usize on a 32 bit target) is not
big enough to support the full range of valid client-allocated object
ids's.