refactor(build): change generated name for protocol
The generated name for the enum that implements core::Protocol had been
based on the protocol name (i.e. WaylandProtocol) but this is redundant
because the protocol enum is already in a module named after the
protocol name. Change the name of the generated protocol enum to be the
literal name Protocol.