feat(client): Add Registry type
A Registry is a key-value store for Wayland protocol global object
records that can produce a stream of change notifications. Registry is
intended to be used by the client code to track the available Wayland
protocol global objects for the purpose of implementing the wl_registry
protocol object. It may also be extensible for use in the server code as
well.