refactor(core): add id method to three traits
Add id method to Interface, Protocol, and ProtocolFamily. This allows
retrieving the ObjectId for a live protocol object once it has been
created since the protocol object is reprsented as a instance of a type
the implements ProtocolFamily (which is expect to be an enum or other
sum type).

The code generated for the id method on ProtocolFamily calls into the
implementations on Protocol which in turn call into the implementations
on Interface. The types that implement Interface are expected to store
their ObjectId internally.