refactor(build): silence dead_code lints in model
The model has several fields that are necessary to properly parse a
protocol XML file but that we don't (yet) use for generating code.
Silence dead_code lints related to the filed by either an
allow(dead_code) attribute on the containing struct (if the whole struct
is not used) or renaming the fields to start with and underscore and
adding a serde(rename) attribute the the field (for unused fields in an
otherwise used struct).