feat(client): implement remaining registry support (#8)
* feat(registry): add iteration of the registry

The iteration happens under a read lock and returns items of type
(&u32, &Global) where the u32 is the numberic name of the global.

* feat(client): add iteration of the registry

The registry method on client::Display allows iteration of the current
contents of the registry and locks the registry during iteration to
prevent concurrent changes.

* feat(client): add wl_registry::global support

Add support for dispatching asychronously recieved wl_registry::global
events that arise after the initial handshake.

* style: run rustfmt

* feat(client): add wl_registry::global_remove support

Add support for dispatching asynchronously received
wl_registry::global_remove events.

* doc(registry): fix spelling and style of a doc comment

* refactor(registry): add getters to Global

The getters for interface and verison in Global allow the type for
interface to be &CStr, instead of CString. This also allows the
implementation of Global to be encapsulated in the registry module.

* style(client): removed unneended todo markers

* style(registry): remove unneeded todo markers

* refactor(client): extract dispatch handlers

Extract separate dispatch handler function for the events of wl_display,
wl_callback, and wl_registry.

* feat(client): add registry CancelationToken

The registry CancelationToken is used by the registry for backpressure.
It causes the dispatch to fail with a RegistryBackpressure error. This
will be used to support a (not-yet-implemented) API to observe changes
to the registry.

* feat(client): implement common traits for API

Add Debug to all public API types and Clone, Copy, PartialEq, Eq,
PartialOrd, Ord, and Hash where they make sense.

* feat(client): add registry change notification

The Display::registry_changes method returns a Stream of
RegistryChange's (Add and Remove).

* style(registry): remove unneed #[allow(dead_code)]

* doc(client): fix spelling errors

* doc(client): fix wording for RegistryRef docs